import copies of TextArea's supporting classes
was originally going for a simple, innocent subclassing of TextArea.
But TextAreaSkin's scrollPane field was private,
preventing me from modifying it (killing it with fire).

So this is the collection of supporting classes needed to have a copy
of JavaFx's TextAreaSkin class in this repo.

All the files have been tweaked slightly so they compile:
referencing each other instead of javafx where appropriate,
package declaration updates,
renaming TextArea to TextAria.

But other than that, they are unmodified (yet).