partially implement TextAria expanding to fit rows of wrapped text
currently only sets the hieght properly after the first relayout,
which is triggered by clicking on the TexArea.
But this is significant progress after a lot of unsuccessful attempts.

this commit includes debug printouts and
unused (unsuccessful) implementations,
just so they are stored in git.

TODO: force a relayout in the first instance.